add isl_map_get_range_simple_fixed_box_hull
This code is essentially copied from PPCG.
It computes an overapproximation of the range of a map
in the form of a fixed-size rectangular box.
It may be a bit PPCG specific, but there is at least one other
potential user (which is arguably closely related to PPCG).
The exact position and size of the box may depend on the internal
representation of the map and is therefore not defined exactly.
This is similar to some of the other hull operations.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>